requirements.txt
# 视频/音频处理
ffmpeg-python==0.2.0
pydub==0.25.1

# ASR (Whisper)
openai-whisper==20231117
torch>=2.0.0
torchaudio>=2.0.0

# AI 服务
openai>=1.0.0

# HTTP 客户端
requests>=2.31.0

# 配置管理
pyyaml>=6.0
python-dotenv>=1.0.0

# 数据处理
pydantic>=2.0.0

# 日志
loguru>=0.7.0

# 工具
tqdm>=4.65.0
